Install pre-commit as a development dependency
This package is used in the following to set up git pre-commit hooks to
ensure the code quality of this project.

In concrete, pre-commit will be used to set up hooks for various code
quality tools. If any of these code quality tools detects an issue with
the code that is about to be committed, it rejects the commit. Thus, the
developer first has to fix the code issues before being able to
successfully commit to his or her local repository.